requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.1.0
certifi @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_4d2o0arbv6/croot/certifi_1665076671099/work/certifi
charset-normalizer==2.1.1
frozenlist==1.3.3
git-python==1.0.3
gitdb==4.0.10
GitPython==3.1.29
humanfriendly==10.0
idna==3.4
multidict==6.0.3
nextcord==2.3.2
nextcord-ext-help-commands==0.0.1
psutil==5.9.4
python-dotenv==0.21.0
requests==2.28.1
smmap==5.0.0
typing_extensions==4.4.0
urllib3==1.26.13
wavelink==1.3.4
yarl==1.8.2
